Made AllEqual more flexible in number of args...
However, this does mean that the previously acceptable
AllEqual("compare", stringSlice...) is no longer valid and should
instead be written, as an example,
AllEqual(append(stringSlice, "compare")...).

My gut (and number of forks and reported imports on godoc.org) tells me
this shouldn't be an issue; but if it really is for you I can revert
this commit.